+++ src/Elders.Pandora.Api/Startup.cs
@@ -31,9 +31,6 @@ namespace Elders.Pandora.Api
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
-            loggerFactory.AddConsole(Configuration.GetSection("Logging"));
-            loggerFactory.AddDebug();
-
             app.UseIISPlatformHandler();
 
             ApplicationConfiguration.SetContext("Elders.Pandora.Api");