+++ src/Elders.Pandora/Pandora.cs
@@ -88,29 +88,5 @@ namespace Elders.Pandora
 
             return merged;
         }
-
-    }
-
-    public class PandoraOptions
-    {
-        public PandoraOptions()
-        {
-
-        }
-
-        public PandoraOptions(string clusterName, string machineName, bool useRawSettingsNames)
-        {
-            ClusterName = clusterName;
-            MachineName = machineName;
-            UseRawSettingsNames = useRawSettingsNames;
-        }
-
-        public string ClusterName { get; set; }
-
-        public string MachineName { get; set; }
-
-        public bool UseRawSettingsNames { get; set; }
-
-        public static PandoraOptions Defaults = new PandoraOptions() { ClusterName = "local", MachineName = Environment.MachineName, UseRawSettingsNames = false };
     }
 }