+++ src/Elders.Pandora.Cli/RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+#### 0.4.0-beta0001 - 25.01.2016
+* Now using PandoraOptions to open Pandora.Box
+
 #### 0.1.1 - 29.05.2015
 * Make the Pandora.Cli a separate package